Refactor FAQ data structure: move FAQ items and categories to a separate mock file; update imports in FAQ components
