+++ src/types/faqs/faq.types.ts
@@ -1,0 +1,14 @@
+export type FAQCategory =
+  | 'tramites'
+  | 'soporte'
+  | 'contactos'
+  | 'historia'
+  | 'gobierno';
+
+export interface FAQItem {
+  id: string;
+  question: string;
+  answer: string;
+  category: FAQCategory;
+  updated_at: Date;
+}
